+++ src/log_formatter_text.erl
@@ -97,4 +97,14 @@ format_metadata(Metadata) ->
 
 -spec format_metadata_pair({atom(), term()}) -> unicode:chardata().
 format_metadata_pair({Name, Value}) ->
-  [atom_to_binary(Name), $=, io_lib:format(<<"~0tp">>, [Value])].
+  [atom_to_binary(Name), $=, format_metadata_value(Value)].
+
+-spec format_metadata_value(term()) -> unicode:chardata().
+format_metadata_value(Value) when is_binary(Value) ->
+  quote_string(Value);
+format_metadata_value(Value) ->
+  io_lib:format(<<"~0tp">>, [Value]).
+
+-spec quote_string(binary()) -> unicode:chardata().
+quote_string(String) ->
+  json:serialize(String).